add zshrc

+++ setup.sh
@@ -4,32 +4,27 @@
 if !(type 'brew' > /dev/null 2>&1); then
     xcode-select --install
     /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+    brew tap caskroom/cask
 fi	
 
 #"===== install nightly-itrem2 ====="
-if [ -e /Applications/iTerm.app ]; then
-    :
-else
+if !([ -e /Applications/iTerm.app ]); then
     open -a 'Safari' "https://www.iterm2.com/nightly/latest"
     https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/Hack.zip
 fi
 
 #"===== dein install ====="
-if [ -e ~/.vim/dein/repos/github.com/Shougo/dein.vim ]; then
-    :
-else
+if !([ -e ~/.vim/dein/repos/github.com/Shougo/dein.vim ]); then
     mkdir -p ~/.vim/dein/repos/github.com/Shougo/dein.vim
-    git clone https://github.com/Shougo/dein.vim.git     ~/.vim/dein/repos/github.com/Shougo/dein.vim
+    git clone https://github.com/Shougo/dein.vim.git ~/.vim/dein/repos/github.com/Shougo/dein.vim
     echo 'dein.vim install'
 fi
 
 #"===== make dotfiles alias ====="
-DOT_FILES=(.vimrc .gitconfig .bash_profile .zshrc)
+DOT_FILES=(.vimrc .gitconfig .bash_profile .zshrc .latexmkrc)
 for file in ${DOT_FILES[@]};
 do
-    if [ -e $HOME/$file ]; then
-        :
-    else
+    if !([ -e $HOME/$file ]); then
         ln -s $HOME/dotfiles/$file $HOME/$file
     fi
 done
@@ -38,10 +33,8 @@ done
 VIM_OPTIONS=(rc colors)
 for option in ${VIM_OPTIONS[@]};
 do
-    if [ -e $HOME/.vim/$option/ ]; then
-        :
-    else
-        mkdir $HOME/.vim/$option
+    if !([ -e $HOME/.vim/$option/ ]); then
+         mkdir $HOME/.vim/$option
     fi
 
     for file in $(ls $HOME/dotfiles/.vim/$option);
@@ -50,19 +43,23 @@ do
         done
 done
 
-#"===== dock setting=====
-defaults write com.apple.dock autohide-time-modifier -int 0
-
 #"===== install python =====
-brew install pyenv python3
+if !(type 'python3' > /dev/null 2>&1); then
+    brew install pyenv python3
+    pip3 install pipenv flake8
+fi	
 
 #"===== setting for tex =====
-brew tap caskroom/cask
-brew cask install skim
-#sudo tlmgr install latexmk
+if !(type 'tex' > /dev/null 2>&1); then
+    brew install ghostscript
+    brew cask install basictex skim 
+fi
 
 #"===== startup apps =====
-brew cask install clipy alfred coteditor insomniax touchswitcher
+if !([ -e /Applications/clipy.app ]); then
+    brew cask install clipy alfred coteditor insomniax touchswitcher
+    open -a 'Safari' "https://www.jetbrains.com/idea/download/download-thanks.html?code=IIC"
+fi
 
 #"===== setting for matplotlib =====
 mkdir $HOME/.matplotlib
